examples/cryptoguard/pyproject.toml
[project]
name = "cryptoguard"
version = "0.1.0"
description = "VLM-based cryptocurrency scam detection in short-form videos"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "Pillow>=10.0",
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
    "torch>=2.1",
    "transformers>=4.45",
    "accelerate>=0.30",
    "yt-dlp>=2024.1",
    "httpx>=0.27",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.setuptools.packages.find]
include = ["cryptoguard*"]
